fix picture endpoint

+++ app/controllers/api/v1/pictures_controller.rb
@@ -2,13 +2,20 @@ module Api
   module V1
     class PicturesController < ApiController
       def destroy
-        picture = current_user.gallery.pictures.find(params[:id])
-        begin
+        picture = current_user.gallery.pictures.find_by(id: params[:id])
+
+        if picture
           picture.destroy
-          render json: { status: 'success' }, status: :ok
-        rescue StandardError => e
-          render json: { errors: e }, status: :unprocessable_entity
+          render json: current_user, status: :ok
+        else
+          render json: { errors: 'Η φωτογραφία δεν υπάρχει' }, status: :unprocessable_entity
         end
+      end
+
+      private
+
+      def picture_params
+        params.permit(:id)
       end
     end
   end
